+++ android/app/src/main/java/io/coursepick/coursepick/data/SearchPlacesDto.kt
@@ -1,9 +1,9 @@
 package io.coursepick.coursepick.data
 
-import io.coursepick.coursepick.domain.Coordinate
-import io.coursepick.coursepick.domain.Latitude
-import io.coursepick.coursepick.domain.Longitude
-import io.coursepick.coursepick.domain.Place
+import io.coursepick.coursepick.domain.course.Coordinate
+import io.coursepick.coursepick.domain.course.Latitude
+import io.coursepick.coursepick.domain.course.Longitude
+import io.coursepick.coursepick.domain.search.Place
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
