+++ android/app/src/main/java/io/coursepick/coursepick/domain/course/InclineType.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.domain
+package io.coursepick.coursepick.domain.course
 
 enum class InclineType {
     UPHILL,
