+++ android/app/src/main/java/io/coursepick/coursepick/domain/course/Length.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.domain
+package io.coursepick.coursepick.domain.course
 
 @JvmInline
 value class Length(
