+++ android/app/src/main/java/io/coursepick/coursepick/domain/course/Segment.kt
@@ -1,4 +1,4 @@
-package io.coursepick.coursepick.domain
+package io.coursepick.coursepick.domain.course
 
 data class Segment(
     val inclineType: InclineType,
