+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesUiEvent.kt
@@ -1,6 +1,6 @@
 package io.coursepick.coursepick.presentation.course
 
-import io.coursepick.coursepick.domain.Coordinate
+import io.coursepick.coursepick.domain.course.Coordinate
 
 sealed interface CoursesUiEvent {
     class FetchCourseSuccess(
