+++ android/app/src/main/java/io/coursepick/coursepick/presentation/search/SearchViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import io.coursepick.coursepick.databinding.ItemPlaceBinding
-import io.coursepick.coursepick.domain.Place
+import io.coursepick.coursepick.domain.search.Place
 
 class SearchViewHolder private constructor(
     private val binding: ItemPlaceBinding,
