+++ android/app/src/main/res/layout/item_place.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="place"
-            type="io.coursepick.coursepick.domain.Place" />
+            type="io.coursepick.coursepick.domain.search.Place" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
