+++ android/app/src/test/java/io/coursepick/coursepick/domain/CourseNameTest.kt
@@ -1,5 +1,6 @@
 package io.coursepick.coursepick.domain
 
+import io.coursepick.coursepick.domain.course.CourseName
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
