+++ android/app/src/test/java/io/coursepick/coursepick/domain/CourseTest.kt
@@ -1,5 +1,9 @@
 package io.coursepick.coursepick.domain
 
+import io.coursepick.coursepick.domain.course.Course
+import io.coursepick.coursepick.domain.course.CourseName
+import io.coursepick.coursepick.domain.course.Distance
+import io.coursepick.coursepick.domain.course.Length
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
