+++ android/app/src/test/java/io/coursepick/coursepick/domain/SegmentTest.kt
@@ -1,5 +1,10 @@
 package io.coursepick.coursepick.domain
 
+import io.coursepick.coursepick.domain.course.Coordinate
+import io.coursepick.coursepick.domain.course.InclineType
+import io.coursepick.coursepick.domain.course.Latitude
+import io.coursepick.coursepick.domain.course.Longitude
+import io.coursepick.coursepick.domain.course.Segment
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
