+++ android/app/src/test/java/io/coursepick/coursepick/domain/fixture/CoordinateFixture.kt
@@ -1,5 +1,5 @@
 package io.coursepick.coursepick.domain.fixture
 
-import io.coursepick.coursepick.domain.Coordinate
+import io.coursepick.coursepick.domain.course.Coordinate
 
 val COORDINATE_FIXTURE = Coordinate(LATITUDE_FIXTURE, LONGITUDE_FIXTURE)
